+++ pkg/network/_examples/broadcast/main.go
@@ -53,9 +53,10 @@ func main() {
 }
 
 func handlerPingPong(serviceName string) network.IHandlerF {
-	return func(n network.INode, _ conn.IConn, reqBytes []byte) {
+	return func(n network.INode, _ conn.IConn, reqBytes []byte) error {
 		defer n.BroadcastPayload(payload.NewPayload(serviceHeader, reqBytes))
 		fmt.Printf("service '%s' got '%s'\n", serviceName, string(reqBytes))
+		return nil
 	}
 }