+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### IMPROVEMENTS
 
 - Update `cmd/hidden_lake/messenger`: append _daemon scripts
+- Update `pkg/network`: append return error for IHandlerF
+- Update `cmd/hidden_lake/traffic`: now HLT redirect messages from producers to network/consumers
 
 ### CHANGES